requirements.txt
torch>=2.7.1
torchvision>=0.22.1
diffusers>=0.34.0
transformers>=4.54.1
accelerate>=1.9.0
xformers>=0.0.31.post1
opencv-python>=4.12.0.88
numpy>=2.3.2
Pillow>=11.3.0
matplotlib>=3.10.5
tqdm>=4.67.1
wandb>=0.21.0
omegaconf>=2.3.0
einops>=0.8.1
controlnet-aux>=0.0.10
albumentations>=2.0.8